Provide rudimentary incremental updates
This is a very crude way of testing whether a file needs updating (in
which case it is fetch completely anew), or not: we just parse the
file and count the version numbers between "head\t..." and "desc".
If the count of version numbers is right, we just skip the file.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>