fix to a segfault, from: fantasai dot lists at inkedblade dot net
Found in a message dated 2009-02-03 16:03:16 PST at:
 http://cvs2svn.tigris.org/ds/viewMessage.do?dsForumId=1670&dsMessageId=1099328
with subject: cvsclone.l bug.

-- Full message body: --

Just finished converting a remote CVS repository to SVN with cvsclone. I did run into a problem, where the particular repository I was working against seems to serve up a different rlog format than cvsclone expects, resulting in a lot of "Unmatched" errors and a segmentation fault (segfault, in case anyone's searching for this problem).


In my case, the problem was an unexpected semicolon after the number of changed lines field. I changed this line in cvsclone.l:

<ATR1>-[0-9]+\n BEGIN REV3; rfile.revt->ldel = atoi(yytext+1);

to this:

<ATR1>-[0-9]+;?\n BEGIN REV3; rfile.revt->ldel = atoi(yytext+1);

and that seems to have fixed the problem. I don't know flex at all, so perhaps that wasn't the best fix, but it worked, and the repository seems to have copied over and imported to svn ok.

Just fyi.